+++ areas/storage/src/AzureMcp.Storage/Commands/StorageJsonContext.cs
@@ -15,6 +15,7 @@ using AzureMcp.Storage.Commands.Table;
 namespace AzureMcp.Storage.Commands;
 
 [JsonSerializable(typeof(BlobListCommand.BlobListCommandResult))]
+[JsonSerializable(typeof(BlobDetailsCommand.BlobDetailsCommandResult))]
 [JsonSerializable(typeof(BatchSetTierCommand.BatchSetTierCommandResult))]
 [JsonSerializable(typeof(AccountListCommand.AccountListCommandResult), TypeInfoPropertyName = "AccountListCommandResult")]
 [JsonSerializable(typeof(AccountCreateCommand.AccountCreateCommandResult), TypeInfoPropertyName = "AccountCreateCommandResult")]